New bootstrapping method based on different build directories

+++ common/lib/mem.c
@@ -36,8 +36,8 @@ void __memset(char *ptr, char value, size_t num) {
 }
 
 void *memmove(void *dest, const void *src, size_t num) {
-  unsigned int udest = (unsigned int)dest;
-  unsigned int usrc = (unsigned int)src;
+  unsigned long udest = (unsigned int)dest;
+  unsigned long usrc = (unsigned int)src;
   if (udest <= usrc)
     return memcpy(dest, src, num);
